requirement.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.3
django-background-tasks==1.2.0
django-compat==1.0.15
future==0.17.1
idna==2.8
pkg-resources==0.0.0
pycparser==2.19
pycryptodome==3.9.0
python-decouple==3.1
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
